requirements.txt
pyzotero
requests
numpy
pandas
slackclient
pylint
flask
slackeventsapi
gunicorn


